Rename io.bitsquare.{=>gui}.FatalException
Localize this exception within the gui package for now, as it's the only
place where it's used at the moment.
